Disallow trailing comma in JSON files
Motivation:

We observed an unexpected deserialization outcome with malformed JSON.
Jackson does not fail on that data but silently drops part of the data.

For example, we expected Jackson to fail to parse the data.
```json5
{
  "a" : {},
  "b" : {},
  "c" : {
    },
    "c1" : true
  }  // a malformed JSON
  "d" : {},
  "e": {}
}
```
But Jackson just drops `d` and `e` fields and parses the JSON into:
```
{
  "a":{},
  "b":{},
  "c":{},
  "c1":true
}
```

The maintainer of Jackson said it is an expected behavior but not a bug.
https://github.com/FasterXML/jackson-core/issues/808#issuecomment-1211057640
> But this is intended behavior, not a bug: Jackson only reads as much
> content as it needs for a valid JSON Value. It does not try to consume
> everything there might be available, given that the white-space-separated
> content streams are commonly used.

He suggested enabling `DeserializationFeature.FAIL_ON_TRAILING_TOKENS`
to fail on the malformed data added after the end of the top-level JSON.

Modifications:

- Enable `DeserializationFeature.FAIL_ON_TRAILING_TOKENS` to the default
  Jackson `ObjectMapper` in Central Dogma
- Breaking) Central Dogma no longer allow tailing commas in JSON

Result:

Central Dogma now correctly rejects malformed JSON when parsing JSON
data.