Allow including raw content in JSON `Change`
Motivation:

As a preliminary work for #1215, add code in advance to maintain
compatibility when JSON is provided as a raw content in `Change`.
If a new version of the client sends `rawContent` in `Change` but the
server hasn't been updated yet, the `rawContent` will be nomalized in
the same way as the current `content` then stored.

Modifications:

- Parse `rawContent` into `JsonNode` if exists and set it to
  `content` for forward compatibility.

Result:

Implemented preliminary compatility work for JSON 5 support