Support for JSON5 (#1222)
Motivation: 

This PR enables JSON5 features available in Jackson, preserves JSON5
content as-is, and still provide the same support such as JSON Path that
was available for JSON.

Limitation: 

Comments in JSON5 content will be lost when a JSON Patch operation is
applied. If you want to preserve the original content, you must use
`UPSERT_JSON`.

Related: #538

Modifications:

- Java implementations)
  - `Json5` is introduced to deserialize JSON5 content.
- Hexadecimal integers (0xdecaf) and trailing decimal points (8675309.)
are unsupported due to the limitation of Jackson.
    - Thanks to @ks-yim, most of the code comes from #655
- `Json5` is used in `Entry` and `JsonChange` to convert JSON5 file into
`JsonNode`.
- Updated `Util.JSON_FILE_PATH_PATTERN` to allow the `.json5` extension.
- `CentralDogmaRepository.importDir()` now uploads raw JSON files
without normalization.
- Fixed `AbstractGitMirror` to mirror raw JSON content instead of the
normalized one.
- Webapp implementations)
- Implemented JSON5 tokenizer for Monaco editor to provide syntax
highlighting.
    - It aligns Jackson's JSON5 features.
  - Add `json5` dependency to validate JSON5 content.
- Fix `NewFile.tsx` to extract a file extension and set it to `language`
for Monaco editor.
- Previously, it was hardcoded to `json` so other languages such as YAML
were not supported.


Result:

- Central Dogma now provides first-class support for JSON5.
- Closes #538

---------

Co-authored-by: KS. Yim <point1304@gmail.com>